The changes here enables a new monitor mode in sentinel to monitor read-only relay masters in cross-datacenter replication setups where there are read-only relay masters replicating from writeable masters in other datacenters and the relay masters have slaves connected to them locally for reads scaling. When such relay masters becomes unavailable, we want the slaves to be switched to the other relay master in the local colo instead of making them sync from a cross-colo master.
Introduced a new sentinel config
'sentinel candidate-slave master1 the-other-relay-master-hostname port'
The stock sentinel failover chooses one of the connected slaves of the
current master and promotes it as master during failover. this new
config
specifies the explicit candidate slave to be chosen as new master for
the connected slaves of the current master. Also, if this config is set,
sentinel does not send 'slaveof no one' to the promoted slave, as the
promoted slave would be an another local relay master replicating from a
other colo master.
A relay master is deemed down when it is either:
    - not replying to ping, or
    - master_link_status of relay master is down
and the other sentinels monitoring the same relay master also agrees on
the above conditions.